refactor(pipeline/scrape): extract ContainerPicker module
Phase 5 — Commit 7 of 9 (pipeline-decoupling-master-2026-05-28).

Carves the transaction-container picking concern out of
ScrapeAutoMapper.ts into a dedicated sub-module under the
per-cluster max-lines:150 eff cap (56 eff).

ContainerPicker/ContainerPicker.ts (56 eff) owns:

- extractTransactions — the workhorse extractor. Walks the
  response body with `huntTransactions`, drops voided rows via
  `isVoidedTransaction`, maps the survivors with
  `autoMapTransaction`, and emits a debug summary count.
- findIndexedSubtree — Isracard / Amex `Index{cardId}` subtree
  lookup. Pure key probe — STRICT_SCOPE step 1.
- filterByCardIndex — value-BFS fallback that filters items by
  the `cardIndex` field. STRICT_SCOPE step 2.
- extractTransactionsForCard — composes the 3-step
  STRICT_SCOPE anti-mirroring chain: Index{cardId} subtree →
  cardIndex value BFS → empty (no fallback). The empty-return
  on step 3 is deliberate: returning the full extraction here
  would mirror every card's transactions onto every other card
  in the same statement window, which is the historical
  Isracard/Amex bug this card-aware extraction layer exists to
  prevent. The warn-level log path is preserved so operators
  see the per-card miss without the run silently double-billing.

ScrapeAutoMapper.ts goes from 214 → 159 eff LoC (-55). The
public re-export block at the bottom continues to surface every
symbol every consumer pulls today; the bottom is now organised
as one `export from` block per sub-module so the dependency map
is readable at a glance.

`getDebug`/`LOG` are removed from ScrapeAutoMapper — every
remaining function in the file is a pure transformer that does
not log; logging stays inside the sub-modules whose
responsibility it is.

No production behaviour change. The container-picker debug-log
`module` string now reads `container-picker` instead of
`scrape-auto-mapper`, but no test asserts on that string
(grep'd src/Tests for `scrape-auto-mapper` returns no matches).
All 66 ScrapeAutoMapper-related tests still pass unchanged.
Husky 15-gate hook green.
