+++ src/Scrapers/Pipeline/Mediator/Scrape/ContainerPicker/ContainerPicker.ts
@@ -1,0 +1,106 @@
+/**
+ * Container picker — extracts transactions from API responses with
+ * an optional card-aware 3-step resolution chain. Owns the
+ * `extractTransactions` workhorse plus the
+ * `extractTransactionsForCard` STRICT_SCOPE anti-mirroring chain
+ * (Index{cardId} subtree → cardIndex value BFS → empty).
+ *
+ * Extracted from ScrapeAutoMapper as part of the Phase 5
+ * pipeline-decoupling split (master plan
+ * pipeline-decoupling-master-2026-05-28 / phase-5).
+ */
+
+import type { ITransaction } from '../../../../../Transactions.js';
+import { getDebug } from '../../../Types/Debug.js';
+import type { ApiRecord } from '../AutoMapperFacade/AutoMapperTypes.js';
+import { isSearchableObject } from '../BfsFieldSearch/BfsFieldSearch.js';
+import { castSearchable } from '../BfsFieldSearch/TxnSignature.js';
+import findFirstArray from '../FieldHunt/LifoCrawl.js';
+import huntTransactions from '../FieldHunt/TxnHunt.js';
+import { autoMapTransaction, isVoidedTransaction } from '../TxnMapper/TxnMapper.js';
+
+const LOG = getDebug(import.meta.url);
+
+/**
+ * Extract transactions from an API response using stack-based
+ * iterative hunt. Filters voided/summary rows. Maps to ITransaction.
+ * @param responseBody - Parsed JSON response body.
+ * @returns Array of mapped ITransactions.
+ */
+function extractTransactions(responseBody: ApiRecord): readonly ITransaction[] {
+  const items = huntTransactions(responseBody);
+  const valid = items.filter((r): boolean => !isVoidedTransaction(r));
+  const mapped = valid.map(autoMapTransaction);
+  const kept = mapped.filter((t): t is ITransaction => t !== false);
+  const count = String(items.length);
+  const validCount = String(valid.length);
+  const keptCount = String(kept.length);
+  const msg = `huntTransactions: ${count} found, ${validCount} valid, ${keptCount} mapped`;
+  LOG.debug({ message: msg });
+  return kept;
+}
+
+/**
+ * Step 1: Key-based lookup — find `Index{cardId}` subtree in
+ * response. Isracard/Amex pattern:
+ * `CardsTransactionsListBean.Index0`, `.Index1`, etc.
+ * @param body - API response body.
+ * @param cardId - Card index (e.g. '0', '1', '5').
+ * @returns Subtree record if found, false otherwise.
+ */
+function findIndexedSubtree(body: ApiRecord, cardId: string): ApiRecord | false {
+  const indexKey = `Index${cardId}`;
+  const values = Object.values(body);
+  const nested = values.filter((v): boolean => isSearchableObject(v));
+  const records = nested.map((v): ApiRecord => v as ApiRecord);
+  const match = records.find((rec): boolean => indexKey in rec);
+  if (match) return match[indexKey] as ApiRecord;
+  return false;
+}
+
+/**
+ * Step 2: Value-based BFS — filter transaction items by cardIndex
+ * field.
+ * @param body - API response body.
+ * @param cardId - Card index to match.
+ * @returns Filtered transaction items, empty if none matched.
+ */
+function filterByCardIndex(body: ApiRecord, cardId: string): readonly ITransaction[] {
+  const allItems = findFirstArray(body);
+  const searchable = castSearchable(allItems);
+  const matched = searchable.filter((item): boolean => String(item.cardIndex) === cardId);
+  if (matched.length === 0) return [];
+  const mapped = matched.map(autoMapTransaction);
+  return mapped.filter((t): t is ITransaction => t !== false);
+}
+
+/**
+ * Card-aware extraction — 3-step resolution chain.
+ * 1. Key lookup: `Index{cardId}` subtree (Isracard/Amex)
+ * 2. Value BFS: filter by `cardIndex` field value
+ * 3. STRICT_SCOPE: empty (no fallback — prevents cross-card
+ *    mirroring that would attribute every card's txns to every
+ *    other card in the same statement window).
+ * @param body - API response body.
+ * @param cardId - Card index for scoping.
+ * @returns Transactions for the specified card only.
+ */
+function extractTransactionsForCard(body: ApiRecord, cardId: string): readonly ITransaction[] {
+  const subtree = findIndexedSubtree(body, cardId);
+  if (subtree) {
+    LOG.debug({ message: `extractForCard: Index${cardId} → key lookup` });
+    return extractTransactions(subtree);
+  }
+  const byValue = filterByCardIndex(body, cardId);
+  if (byValue.length > 0) {
+    const count = String(byValue.length);
+    LOG.debug({ message: `extractForCard: cardIndex=${cardId} → value BFS (${count} txns)` });
+    return byValue;
+  }
+  LOG.warn({
+    message: `STRICT_SCOPE: no data for Card ${cardId} — returning empty (no fallback)`,
+  });
+  return [];
+}
+
+export { extractTransactions, extractTransactionsForCard };
